+++ spec/integration/client/client_spec.rb
@@ -239,4 +239,44 @@ EOM
     end
 
   end
+
+  when_the_repository "has a cookbook with only an audit recipe" do
+
+    before do
+      file 'config/client.rb', <<EOM
+local_mode true
+cookbook_path "#{path_to('cookbooks')}"
+audit_mode :enabled
+EOM
+    end
+
+    it "should exit with a zero code when there is not an audit failure" do
+      file 'cookbooks/audit_test/recipes/succeed.rb', <<-RECIPE
+controls "control group without top level control" do
+  it "should succeed" do
+    expect(2 - 2).to eq(0)
+  end
+end
+      RECIPE
+
+      result = shell_out("#{chef_client} -c \"#{path_to('config/client.rb')}\" -o 'audit_test::succeed'", :cwd => chef_dir)
+      expect(result.error?).to be_falsey
+      expect(result.stdout).to include("Successfully executed all `controls` blocks and contained examples")
+    end
+
+    it "should exit with a non-zero code when there is an audit failure" do
+      file 'cookbooks/audit_test/recipes/fail.rb', <<-RECIPE
+controls "control group without top level control" do
+  it "should fail" do
+    expect(2 - 2).to eq(1)
+  end
+end
+      RECIPE
+
+      result = shell_out("#{chef_client} -c \"#{path_to('config/client.rb')}\" -o 'audit_test::fail'", :cwd => chef_dir)
+      expect(result.error?).to be_truthy
+      expect(result.stdout).to include("Failure/Error: expect(2 - 2).to eq(1)")
+    end
+  end
+
 end